code organization and bug fix
